sanity_check/requirements.txt
datasets>=2.0.0
gensim>=4.0.0
numpy>=1.20.0
scikit-learn>=1.0.0
matplotlib>=3.0.0
umap-learn>=0.5.0
tqdm>=4.62.0
torch>=1.9.0
transformers>=4.10.0
biopython>=1.79
pyarrow>=14.0.0
pandas>=1.3.0
fpdf2>=2.7.0
adjustText 